requirements.txt
streamlit
python-tds
azure-identity
